ARALILA/requirements.txt
asgiref
Django
django-extensions
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
channels
channels_redis
daphne